refactor: use Cheerio only for link discovery, not content extraction
- Rename ContentScraper to LinkDiscoverer
- Remove content extraction functionality from Cheerio implementation
- Pure.md remains the exclusive tool for content extraction
- LinkDiscoverer only extracts links for crawling and pagination
- Pages without successful Pure.md extraction show placeholder content
- Better separation of concerns between content extraction and link discovery
